requirements.txt
pandas
streamlit
altair
numpy
pydeck
plotly.express
